Remove Fly's concurrency hard_limit (#820)
As described in the
[docs](https://fly.io/docs/reference/configuration/#services-concurrency):
> hard_limit : When a Fly Machine is at or over this number of
concurrent connections or requests, the system will stop sending new
traffic to that Machine. If unset, no hard_limit is enforced.
soft_limit : When a Fly Machine is at or over this number of concurrent
connections or requests, the system will deprioritize sending new
traffic to that Machine and only send it to that Machine if all other
Machines are also at or above their soft_limit. This setting is also
used to determine excess capacity for the [auto start and stop
feature](https://fly.io/docs/apps/autostart-stop/). If unset, the
default is 20.

We are keeping `soft_limit` in case we introduce again autoscaling. But
we _must_ remove the `hard_limit` as when the `hard_limit` is reached
and we don't have autoscaling, the server stops accepting any requests,
effectively leading to an outage.

This is confirmed by the latest incident where the response times went
through the roof as we approached the 100 `hard_limit`. Until this PR is
merged, I'm keeping two machines up, effectively increasing the number
of connections we can keep concurrently to 200. Two machines put a
strain on the database, so it's only a temporary measure.


![image](https://github.com/tuist/server/assets/9371695/5627f961-03b5-4572-915e-aac4f1a8110a)