Add new billing page (#823)
This PR adds a new billing page with the business logic to:
- Upgrade and downgrade between Air and Pro plans
- Redirect to a Stripe session if a user doesn't have an active
subscription
- Surface some billing details such as the billing email
- Surface remote cache hits count

The current UI looks as such:

![image](https://github.com/tuist/server/assets/9371695/3d14eb8b-1762-4d7f-9715-b242181b49e7)

There are a couple of follow-ups to this PR which are necessary before
we start shipping this to prod:
- Remove `plan` field -> this can only be done once we migrate all
enterprise customers to the new database model. We will need to assign
each enterprise customer a valid Stripe subscription, so the business
logic can stay more-or-less the same regardless of the plan.
- Whenever a new user is created, we should create a new trial with the
Air plan. This can be done without the user having to fill in their
payment details, see
[here](https://docs.stripe.com/payments/checkout/free-trials#section-1)
for how to do this from the Stripe perspective
- Integrate this with the new layout being built out by @pepicrft

Possible future improvements:
- Surface the current bill for usage
- Consider adding `customers` and `payment_methods` tables, so we don't
have to hit the Stripe API for the billing details.