Change create project API endpoint to accept a full handle (#846)
Counterpart client PR: https://github.com/tuist/tuist/pull/6472

We're changing the `create` project endpoint to accept a `full_handle`
instead of `name` and an optional `organization_name`. While we could
parse the handle on the client, it's better to move as much of the logic
to the client.

I'm keeping the get project with `:account_handle/project_handle` as in
the URLs, it's better to keep the two separated.