Implement new mechanism for user and project tokens (#855)
This PR implements a new token strategy that should make our app
actually secure, rather than us saving plain text tokens in the
database.

Here's the current setup:

### User tokens

For user _access_ tokens, we use stateless JWT tokens, using the
[guardian](https://github.com/ueberauth/guardian) library. These tokens
are short-lived (10 minutes), which is why they can be stateless.

When an access token expires, the client needs to refresh it with their
refresh token. The refresh tokens live for much longer – the current
value is 4 weeks. Since these can be used to access the app
indefinitely, we need to be able to revoke them. That's why we save
these in the database using
[guardian_db](https://github.com/ueberauth/guardian_db). The JWT value
is encrypted, so we never store any tokens in plain text. The refresh
tokens can also be used only once.

### Project tokens

Project tokens are generated without a TTL. Since the project tokens
can't be stateless, it made no sense to use JWT (and thus guardian).
Instead, I created a new table called `project_tokens` to store these.

The process is as follows:
- User creates a new project token via `tuist project tokens create`
- We create a new token in the database. The hash is in the format of
`tuist_uuid-v7-id_crypto-secure-generated-hash`
- The full hash is encrypted in the database
- When using a project token, we check with `Bcrypt` the passed project
token corresponds to the hash in the db (similar to how we store
password)
- There are then to extra commands, `tuist project tokens list` and
`tuist project tokens revoke`. We use the UUID to identify the tokens,
so that users can revoke them (and see the individual tokens)