Fix server-driven warnings causing CLI versions older than 4.11.0 to fail when interacting with the API (#863)
Before 4.11.0 we had a
[bug](https://github.com/tuist/tuist-cloud/commit/ac0e31e89eaaf8e749323d38c4dd28b01b163e7f)
in the client middleware because it assumed the content of the header
was not base64-encoded. We fixed the issue in the client, but from the
server, we wrongly assumed people would be using the fix, which led to
the error arising as soon as we merged the new tokens logic, which makes
use of warnings.