Update usage price IDs for non-prod environments (#867)
I updated the price IDs to be in line with what we have in production. I
couldn't just update those prices as they already had test subscriptions
attached to it and that, unfortunately, means I needed to create a
completely new price.

The specific reason for not being able to update the price is the
following:
> You can only edit the description, lookup key, and tax behavior
because this price was used in a transaction and its tax behavior is not
set. To make other changes, create a new price.

I wonder if in prod, we can ensure the price behavior is set as we will
want to have the possibility to update existing price with new
usage-based items in the future without creating a new price that we'd
need to migrate customers to.