Forwards storage 5xx errors to the client when obtaining the object size (#893)
Fixes https://github.com/tuist/server/issues/892

The storage provider might return 5xx errors. I suggest that we forward
them to the client for retrying. This PR does it only for the function
to get the object size, which is the one for which we got an error
report, but we might need to do it for the others too if those endpoints
turn out to be flaky too.