Not treat any account creation error as "handle taken" error (#902)
While [debugging this
issue](https://github.com/tuist/server/issues/901), I noticed the error
on App Signal because the scenario of an account taken and us retrying
with multiple suffixes is already covered by tests. Looking at the code,
I noticed that we treat all the errors as `account_taken`, which masks
the actual error.
This PR changes that so that we can see the actual error the next time
it happens.