Preparation for making the registry a public repository (#903)
In preparation for making the registry repository public, I require the
`TUIST_LICENSE` to be present for us, too, but in this case, with
`expiration_date` set to `none.` This prevents anyone from running the
service as if it were Tuist-hosted (changes have been validated in
staging).

I also created another Dockerfile for on-premise builds that don't copy
the encrypted secrets. I tried to use a single Dockerfile and
conditionally run a `COPY` step when an `ARG` has a certain value but
that's not supported. Since we don't change the `Dockerfile` frequently,
I think the risk of inconsistencies is quite low.

---------

Co-authored-by: Marek Fořt <marekfort@me.com>