Forward errors starting and completing multipart uploads to the client (#908)
**Fixes:** [Error
1](https://appsignal.com/tuist-cloud/sites/6607b64d83eb6789e61c8ba7/exceptions/incidents/286),
[Error
2](https://appsignal.com/tuist-cloud/sites/6607b64d83eb6789e61c8ba7/exceptions/incidents/272),
and [Error
3](https://appsignal.com/tuist-cloud/sites/6607b64d83eb6789e61c8ba7/exceptions/incidents/290).

R2 intermittently yields 5xx errors that cause the request-response
workflow to abort. I'm adjusting the implementation to forward those to
the client, where we have some retrying logic. The errors coming from R2
are sanitized so I think it should be safe to pass them down to the
client.