Fix missed index due to LOWER function (#924)
Fixes the performance of the following query:
```sql
select
  a0."id",
  a0."name",
  a0."user_id",
  a0."organization_id",
  a0."cache_upload_event_count",
  a0."cache_download_event_count",
  a0."customer_id",
  a0."current_month_remote_cache_hits_count",
  a0."current_month_remote_cache_hits_count_updated_at",
  a0."created_at",
  a0."updated_at"
from
  "accounts" as a0
where
  (lower(a0."name") = $1)
```

The issue of this query is the `lower` method. The index is built with
the column value and it won't be used as soon as you use a method such
as `lower`.

Turns out we don't even have to use `lower` at all since all the columns
already have `citext` type, so the search is case insensitive by
default.

I only needed to change `invitations.invitee_email` type to `citext`,
all the other columns associated with a query that had a `lower` already
have that type.