fix: run detail view not opening due to error parsing the result bundle (#945)
Fixes:
https://appsignal.com/tuist-cloud/sites/6607b64d83eb6789e61c8ba7/exceptions/incidents/273/samples/6607b64d83eb6789e61c8ba7-50221992285151152717234683201

The logic parsing the result bundle makes some assumptions about the
presence of certain attributes. This PR softens that assumption to
mitigate the issue. I couldn't figure out what the absence of the
attribute means in that particular case, so I leaned on mitigating the
error until we can better understand the scenario through users' input.