chore: fix flaky tests caused by Internet hits (#944)
I noticed some flakiness in our tests suite is caused when rendering
live views because the layout, which includes the badge, hits the GitHub
API to get the latest release. Since we don't want to be stubbing that
everywhere, I decided to extract that logic into a `GenServer` that
holds the state in memory and synchronizes periodically. This
`GenServer` is only included in the supervision tree when the
environment is not `test`, and the interface returns `nil` when fetching
the latest version if the server is not running.