fix: genserver timeout trying to get the latest CLI release (#952)
[Error](https://appsignal.com/tuist-cloud/sites/6607b64d83eb6789e61c8ba7/exceptions/incidents/24/samples/6607b64d83eb6789e61c8ba7-1834411523263864803117236782601)

When I created the genserver to persist the state of the latest CLI
release in memory, I used Horder to share it across all the future nodes
if we have more than one instance of the app running. This caused
[issues](https://appsignal.com/tuist-cloud/sites/6607b64d83eb6789e61c8ba7/exceptions/incidents/24/samples/6607b64d83eb6789e61c8ba7-1834411523263864803117236782601).
Since the state we are keeping in memory is not large nor can present
API-limit issues when fetching it from multiple nodes, I think it's safe
to add it as a child of the application when starting it. Hopefully this
will mitigate the above issue.