Add support for updating the connected git repository (#976)
See the counterpart client PR for more context:
https://github.com/tuist/tuist/pull/6671

This PR adds `repository_url` to the `Project` model and adds the
possibility to update it via the REST API.

When updating to a given `repository_url`, we make sure that the actor
has write permissions to the repository. This ensures malicious actors
can't connect a repository they don't have access to.

When the `repository_url` can be updated, we also query the
`default_branch` and update the value based on the response from the
GitHub API. When a git repository is connected, the `default_branch`
should be synced using a webhook.

Additionally, we're now checking that the git remote url specified in
the metadata when creating a command event is equal to the
`project.repository_url`. This ensures we don't create comments in
projects the project shouldn't have access to.