Fix git_remote_url_origin mismatch (#984)
I've noticed tuist PRs are missing tuist GitHub app reports even though
the app is installed and we're running the `tuist` latest version.

I went into the db and the `command_events` do include the git metadata.
But the `git_remote_url_origin` has
`https://tuist@github.com/tuist/tuist` – but the URL we set is
`https://github.com/tuist/tuist`. Instead of matching command events
based on the `git_remote_url_origin`, we can use the `project_id`.

I'm removing the `git_remote_url_origin` as a column and instead only
use the value to get the `full_handle` when the event is created to
ensure we don't post the report to unrelated repositories.