Set up Prometheus metrics endpoint for Fly deploy (#987)
Supersedes: https://github.com/tuist/server/pull/983

The superseded PR is not necessary as we don't need to auth the
`/metrics` endpoint if the metrics endpoint is served on a port that's
not publicly facing:
https://community.fly.io/t/authenticated-requests-for-health-checks-and-prometheus-metric-scraping/7909

I set up a quick dashboard
[here](https://tuist.grafana.net/d/cdx0eoxtsb5dsc/tuist-server),
currently populated with canary data.

The data source is configured
[here](https://tuist.grafana.net/connections/datasources/edit/cdx020i9owlc0b/).
It seems the
[docs](https://fly.io/docs/monitoring/metrics/#authentication) are
somewhat outdated and I wasn't able to create an org readonly token, so
the token in the datasource was obtained by running `flyctl auth token`.
Not ideal and something we might want to change in the future.

Since we changed the port, I also reflected that in the docs for on
premise: https://github.com/tuist/tuist/pull/6703. We _could_
technically make that configurable, but I would only do so once someone
asks for that.