Return not-found page when the command with the given id doesn't exist (#981)
Fixes https://github.com/tuist/server/issues/972

The `ProjectRunDetailLive` view assumes the command event with the id
passed as a param always exists, causing some logic to fail because it
builds on that assumption. I'm adjusting the view to an error with a not
found if the command event doesn't exist.