Use `Req` to interact with AWS (#997)
I noticed some requests to stat the multi-part upload [time
out](https://appsignal.com/tuist-cloud/sites/6607b64d83eb6789e61c8ba7/exceptions/incidents/330).
That's odd since the HTTP client that is used by default, Hackney, is
configured with a timeout of 30 seconds. That makes me think about two
things:

- Cloudflare R2 is slow at responding.
- Hackney is doing something odd, which I'd doubt, considering it's
widespread in the Elixir ecosystem.

This PR replaces Hackney with Req following the recommendation of [this
blog
post](https://andrealeopardi.com/posts/breakdown-of-http-clients-in-elixir/)
of not using Hackney. Since `Req` has a utility to automatically set the
signature headers for S3, I'm using that instead of going through the
ExAws package, which feels like unnecessary indirection.

I've taken the opportunity to report some request events via
`:telemetry` and surface that in the `/metrics` endpoint so that we can
consume it on the Grafana side.

---------

Co-authored-by: fortmarek <marekfort@me.com>