Use http2 and a larger pool size for the HTTP requests against the S3-compliant storage (#1000)
The high CPU consumption and slow requests might be caused by the size
of the current HTTP pool and the usage of HTTP1, which can't take
advantage of multiplexing. This PR:

- Adjusts `:ex_aws` to use `:req`, which internally uses `:finch`, whose
pools are very configurable (we can drop `:ex_aws` in a follow-up PR
because it's not really needed)
- Makes the pool configurable via env. variables (documentation to
follow).