Use transactions to access the cached token validation results (#1015)
It turns out that the previous approach with Nebulex did not handle the
scenario where multiple processes tried to access the cache value while
other processes were regenerating the value. This PR addresses that
using the Cachex package, which has a "transactions" interface similar
to Ecto's for this particular case.