Optimize storing and fetching selective tests (#986)
Right now, we unnecessarily store and retrieve selective test hashes
from the Cloudflare storage – even when they are fingerprint files. To
fix this, I'm adding two new endpoints:
- POST /api/projects/:account_handle/:project_handle/cache/:category
- GET
/api/projects/:account_handle/:project_handle/cache/:category/:hash

The new `POST` endpoint is necessary as we no longer go through the
multipart upload when uploading selective test results. The results are
instead saved in the db in the new `cache_items` table.

The new `GET` endpoint would not be strictly necessary. We could reuse
the existing endpoint and return a different object based on the CLI
version. However, I lean to having a new endpoint where the URL is more
explicit about retrieving a cache item with a specific hash for a
specific project. We can eventually retrieving binaries through that
endpoint by adding a `download_url` field in the `CacheItem` that we
return.

See client PRs for more context:
- https://github.com/tuist/tuist/pull/6701
- https://github.com/tuist/tuist-cloud/pull/264