Move the keygen logic from Rust to Elixir (#1020)
There's no reason to have this logic in Rust when all it does is to send
an HTTP request and cache the value with an invalidation of period of 1
day. The one only reason that we had was obfuscating the validation key
in Rust, but I think it's an unnecessary extra layer of security at this
point.