Remove Rust (#1021)
We moved to Rust to use the AWS Rust SDK, which supports the
authentication mechanism required by one of our on-premise customers.
However, that came at the cost of being unable to reuse HTTP connections
since the Rust executions are stateless. This is very undesirable,
especially considering that sudden bursts of requests can come to the
server, and we might want to have a warmed pool of connections ready to
hit the S3 server.

This PR reverts that by implementing the authentication in Elixir. It
also removes all the Rust code from the repository.