Fix GitHub app authentication (#1022)
Resolves: https://github.com/tuist/tuist/issues/6716

When initially implementing the GitHub app, I was using the
`access_tokens_url` from the
[/installations](https://docs.github.com/en/rest/apps/apps?apiVersion=2022-11-28#list-installations-for-the-authenticated-app)
endpoint. This endpoint returns all installations and we were plucking
out the first one. That, obviously, made things fail when we had more
than one installation.

Instead, we should get the GitHub app installation using the
[/{repo}/installation](https://docs.github.com/en/rest/apps/apps?apiVersion=2022-11-28#get-a-repository-installation-for-the-authenticated-app)
endpoint. To do that, we need to pass the repository's full handle to
all the `client` methods.