Return :ok instead of :bad_request when uploading an already-existing cache action item (#1026)
Currently, when uploading a cache action item that already exists, we
are returning a `:bad_request` response.

However, conflicts _are_ likely to happen because we:
- Fetch hashes
- Run tests for missing hashes
- Upload hashes that were not initially available in the cache

When a cache action item is uploaded while we run tests (which can take
a long time), the upload of those items would fail due to the
`:bad_request`, such as here:
https://codemagic.io/app/65ca555c190cfbe9f5dd792f/build/66f44a0b704a7c906680709c#step:8

Instead of `:bad_request`, I decided to go with `:ok` since we consider
such a request valid. Note there is still difference if an item was
actually created as in that case, we return `:created`.

Alternatives considered:
- `409` Conflict:
https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/409. I
discarded this one because we 1) want to return the cache action item
which would be weird to do in an 4xx response 2) based on the specs, 409
should be returned when resubmitting would help, which is not our case
("are errors sent to the client so that a user might be able to resolve
a conflict and resubmit the request")
- Changing the `POST` request to `PUT` where it's expected that you
would only rewrite the item if it exists.
[This](https://bazel.build/remote/caching#http-caching) is what Bazel is
doing. I _think_ I opted for `POST` as the client action _is_ creating
an item whereas in Bazel's case, it's for uploading a file which is not
what we are doing. I probably would have gone with `PUT` if I did this
from scratch but for now, it's easier to just stay with our current
setup and use `:ok`.

Additionally, I put a fallback to `:created` when the CLI version is
`4.28.0`, so we're not forced to do a new version asap.