Not show billing link in on-premise instances (#1027)
An on-premise organization reported an error in their instance. When
looking at their logs, it turns out that the user tried to access the
billing page of their account, which leads to a bug because the logic
behind the view assumes integration with Stripe.

This PR extends the authorization logic to prevent anyone from reading
the billing information of an account if the environment is on-premise.
Since this authorization logic is used by the layout to show/hide the
billing link, the link will be absent in those environments.

Note that the route is still accessible if you type it in the browser.
However, this is an improbable scenario. If we see it happens (i.e.,
people trying to type the URL), we can do something else:
- Detect that and fail through a plug
- Conditionally remove the route at compile-time, but this comes with
come additional costs for us which I'd rather avoid.