Increase the DB queue target and intervals and make them configurable (#1029)
The best explanation for what those values represent was [this
one](https://elixirforum.com/t/what-does-queue-target-and-queue-interval-exactly-mean-in-dbconnection/33682/3?u=pepicrft).

`queue_interval` is the interval that for checking whether **all** the
connections took more than `queue_target` time to be checked out. If
that happens, it doubles `queue_target` and if all are once again above
the threshold, then it starts dropping.

Because we get bursts of call, a higher queue time is expected, so I
bumped `queue_target` from the default value `50ms` to `300ms`. I also
increased the check interval to capture the period that a burst of
requests can last.

I took the opportunity to make this configuration configurable via
environment variables (documentation to follow).