Handle HTTP and client errors when validating the license in on-premise instances (#1041)
An on-premise user
[reported](https://files.slack.com/files-pri/T04N4ETMVHN-F07PJKEJ2JG/logs-from-tuist-cloud-in-tuist-cloud-8bc8b4d8d-6658r.log)
a startup crash. It's hard to tell what's causing it from the logs, but
looking at the sequence of events, I assume that it's coming from the
validation of the license.

This PR is improving the validation logic to account for HTTP and `:req`
errors and adding some logs. I'll follow-up with them to ensure that
they are setting the license in the environment.