Fix rollback behaviour on failing migration to not rollback migrations that executed successfully (#1042)
One of the on-premise customers reported an issue trying to launch the
app. While looking at the logs, I noticed our `start` script attempted
to roll back not only the migrations that had to run, but the ones that
had already been executed and persisted in the database. It turns out
that the logic was wrong.

We need to use `to_exclusive:` instead of `to:` to roll back down to
that version, which was the one that was persisted in the database,
without running that one.

```elixir
# Note to
Ecto.Migrator.with_repo(repo, &Ecto.Migrator.run(&1, :down, to_exclusive: version))
```

While looking at the `start` script, I also noticed that we roll back,
even in the scenario of the service aborts later on, not just at launch
time. This is also not ideal, because it means the server could abort
after 3 days of running, and they'd end up with an out-of-sync between
the database and the service. I think we should also roll back in the
case a migration fails.