Disable some interactions with the billing APIs from on-premise instances (#1045)
An on-premise customer reported 504 errors in their instance. It turns
out that we were hitting billing APIs from our business logic when
creating users and accounts. This PR fixes that and add some extra tests
for that scenario.
I also removed the `Tuist.Billing.enabled?()` in favor of
`Tuist.Environment.on_premise?()` since I think on-premise already
reflects whether billing should be enabled or not.