Add additional logging to the logic that exchanges the AWS token in a file for the credentials (#1046)
In one of the on-premise customers, requests are receiving a 504 on the
client. I can't tell what might be causing it, so I'm adding some logs
in the logic that resolves the credentials using the AWS token persisted
in a file to check if those requests are hanging for some reason.