Add support for dependencies on app targets from different projects
In certain cases, such as when the test host app resides in a different
project, it's necessary to depend on an .app target from another
project. To enable this, a file reference to the app is required,
followed by setting up a dependency on it. This commit introduces file
references for all .project executable targets, and adds a build phase to
copy the .app file, making it available during the build process.