Improve the email debugging experience in development (#1052)
While helping @asmitbm visualise the emails that we'd send in
development, we noticed that the [Swoosh local
adapter](https://hexdocs.pm/swoosh/Swoosh.Adapters.Local.html), can keep
the emails in memory and present them in a route, which makes them easy
to visualise and debug.

However, the above was not working because we overrode the adapter to
use Mailgun also in development. Moreover, the `/ops/sent_emails`
required authentication, which made seeing emails like the email
confirmation email impossible to debug.

I'm adjusting the configuration to use the local adapter locally, which
comes with the trade-off of distancing ourselves a bit from using the
adapter that we use in production, but I believe it's safe having
validated its correct functioning already. I also changed the
`/ops/sent_emails` to only be present in development without any
authentication or authorization.