Add support for install preview archive on mobile devices (#1053)
Client PR (draft for now): https://github.com/tuist/tuist/pull/6849

This PR implements the server-side functionality for:
https://github.com/tuist/tuist/issues/6847. See this issue for more
background about this feature.

That means:
- Clients can now upload archive (`.ipa`) previews
- `.ipa` previews on desktop are rendered the same way as bundle
previews
- `.ipa` previews on mobile look different. The design as of now is very
basic. I will work with @asmitbm to make it better.
- When clicking on `Install` in `.ipa` previews on mobile, the app is
downloaded on the device

The following is necessary for downloading `.ipa` as per the
[docs](https://support.apple.com/ru-ru/guide/deployment/depce7cefc4d/web):
- A `<a>` link with a specific href:
`itms-services://?action=download-manifest&url={.../manifest.plist}`
- The link leads to a `manifest.plist` that contains basic information
about the app (such as bundle ID) and the download URL for the IPA. The
`manifest.plist` needs to be _public_, accessible by Apple, and hosted
via https. Installing `.ipa` won't work in local development and we need
to use staging.
- A URL for the IPA download

Currently, any `.ipa` preview is _public_ by default, so testers don't
need to be invited to Tuist. In the future, we should allow
organizations to make previews _private_, but some links, such as
`manifest.plist` will need to stay accessible.