Fix email delivery (#1059)
I noticed the email delivery
[broke](https://cloud.tuist.io/ops/errors/23) after [we merged this
PR](https://github.com/tuist/server/pull/1054). It turns out we broke
the contract with the notifier, which after the change expected an
`icon_url` attribute that we were not passing. This PR fixes it by
eliminating the need for that information to be passed, since it's
always the same.