Move authorization logic from the Authentication to the Authorization module (#1057)
As discussed
[here](https://github.com/tuist/server/pull/1053#discussion_r1795515808),
parts of the logic in `Authentication` are more of an authorization. I
updated the logic whether to require the user to be authenticated or not
to use the `Authorization` module.

I decided to use `can(nil, ...)` instead of using the current subject as
using it would technically mean we would run the authorization logic for
a given subject twice. What we need in the `Authentication` module is to
answer the following: "Can an anonymous user access this model?" If the
answer is yes, we don't require the user to be authenticated and vice
versa.