Add support for dependencies on app targets from different projects (#6820)
* Add support for dependencies on app targets from different projects.

In certain cases, such as when the test host app resides in a different
project, it's necessary to depend on an .app target from another
project. To enable this, a file reference to the app is required,
followed by setting up a dependency on it. This commit introduces file
references for all .project executable targets, and adds a build phase to
copy the .app file, making it available during the build process.

* Add executable non-local dependencies fixture and acceptance test.

Add a fixture with executable dependencies (app for iPhone, app for
AppleWatch,  app extension) where the executable target are not part of
the project which depends on them, but rather a part of another project.
Also, add acceptance test for this fixture.

---------

Co-authored-by: Marek Fořt <marekfort@me.com>