Update the token name to be consistent with the 1Password name (#1082)
I noticed weekly releases and deployments are not going through because
the token we were using had no permissions to clone the `cli-ext`
repository to run acceptance tests.
Since we've been in this scenario before, I decided to move the token to
a new GitHub account which I created, `@tuistit` (credentials are on
1Password), and created the token under that user. The token is in
1Password as `TUIST_RELEASE_GITHUB_TOKEN`, and I'm updating the naming
here to be consistent with 1Password. Otherwise it's not trivial to know
which token is being used here.