Add support for getting distinct previews based on bundle identifier (#1087)
In the macOS app, we want to get a list of apps that have previews that
can be associated with the `latest` specifier.

To enable this, we need to be able to tell the previews index route that
we want to get previews with distinct `bundle_identifier` field. This PR
implements that functionality.