Some marketing website improvements (#1099)
- I noticed the hero section of `/blog` didn't have the right and the
left site taking space equally, so I'm fixing it.
- We got some [feedback](https://discuss.systems/@ls/113448289943175318)
on Mastodon on the pricing not being clear. I think people expect us to
show them a fix amount of X per month, which makes me wonder if the idea
of paying for usage is something that will click with customers, or
they'll not use it driven by the fear of being overcharged.