Fix a bug trying to call `category` on a `nil` hero post (#1100)
Fixes
https://appsignal.com/tuist-cloud/sites/6607b64d83eb6789e61c8ba7/exceptions/incidents/320/samples/timestamp/2024-11-08T15:43:00Z

It turns out that the logic tries to obtain the hero post from the list
of filtered posts, which is incorrect. Because of that, we might get a
`nil` post when the selected filter doesn't yield any post, causing the
rendering to fail. This PR fixes it by getting the hero post from the
total list of posts, and assigning it to the socket pre-filtering.