Remove marketing resources and symbols from the image that's built for on-premise customers (#1106)
This PR removes all the marketing sources and resources from the
enterprise image that we build for on-premise customers. To achieve so:
- I've scoped the marketing resources and sources under `marketing/`
directories.
I've adjusted the router to register the marketing routes only when the
`TUIST_HOSTED=1` environment variable is true at compile time.
- I've adjusted automation to pass the right value of that variable as
`--build-arg` to the Docker process that builds the image.
- I've adjusted the `Dockerfile` to remove marketing resources and
sources if the environment variable is falsy.