Gracefully handle scenario when latest App release has no DMG asset (#1111)
Resolves: https://github.com/tuist/server/issues/1110

We currently assume that the latest `app@x.y.z` release has a DMG asset
– that might not always be the case since we:
- Create a GitHub release
- Upload the assets

The upload of assets can fail (which happened during the `app@0.5.0`
release) and then we end up with the latest app release having no DMG
asset. Going to `tuist.dev/download` then throws a 500 error.

When getting the latest app release, we should only consider those that
have a DMG asset. When the upload of assets fails, we'd gracefully
fallback to the previous release.