Compile-time declarative authorization policies (#1072)
This PR takes a stab at using the
[let_me](https://github.com/woylie/let_me) to declare authorization
policies. The package's interface offers several benefits over our
current approach:

- It's declarative at compile-time, which means we can turn the spec
into a standard policy document that we can then read from the
documentation to document the authorization rules.
- It separates declaration from the allow/deny rules, which allows
reusing allow/deny rules across object actions.

> [!NOTE]
> I've only migrated the `:project_cache_read` policy in this PR to get
a sense of the package and assess whether it's a good idea or not for
Tuist. If we agree to make it our default, we can gradually migrate the
others.