Optimize the query that marks test case runs flaky if they are (#1122)
The query that we ran to mark flaky tests as flaky traversed the entire
`test_case_runs` table, traversing even tests whose module hashes were
not within the module hashes of the test case runs that we had just
updated. This PR addresses that by limiting the query to the test case
runs whose module hashes match the ones of the just persisted ones.