Optimize the persistence of test case runs (#1128)
We used to traverse all the projects (N) and execute three queries for
each. This can quickly exhaust the DB connection pool if the project has
a lot of Xcode projects. For example, one of our customers has over 100
projects, so we end up executing 300 queries for an API call. Thanks to
this change, it has been reduced to 3 queries per API hit. Moreover, the
write/update queries are put in a transaction for atomicity and to let
the database perform further optimizations.