Return PNG image instead of SVG in API endpoints (#1127)
<img width="1512" alt="image"
src="https://github.com/user-attachments/assets/a28e0d7d-bbcf-4edf-a8b9-42b6dd84b398">

Adds a new router for `qr-code.png`. I'm keeping the `.svg` variant as
it's more flexible and we use it in the GitHub PR comments. But `.png`
with a reasonable default (right now, we use the library's default which
is `410x410`) is probably better when returning the `qr_code_url` as PNG
is easier to display in multitude of applications, like Slack, rather
than SVG.