Introduce pricing changes (#1125)
This PR introduces the necessary changes to eliminate the $249 base
price from Tuist Pro. Since we can only roll this out with notifying the
users. The access to the changes is gatewayed through
`Tuist.Billing.new?`.

In a follow-up PR, I'll add the logic to deliver the email to notify
accounts about the changes. Reached the day, we'll have to archive the
$249 price item so that it cannot longer be used, and migrate the
existing subscriptions to remove that item.

> [!NOTE]
> I've taken the opportunity to re-use the pricing plans across the
marketing and the dashboard routes