Fix preview icon not found due to different full handle casing (#1148)
When the full handle is written in camel case, such as `IceCubesApp`,
it's saved downcased in the database (`icecubesapp`).

However, when uploading the icon, we use the camelcase variant of the
full handle. When downloading files, Tigris is case _sensitive_, so if
you try to download the icon with a different casing, Tigris won't find
it. I tried to see if the file path can be made case insensitive, but I
couldn't find such a setting.

Alternatively, we could downcase the key at the storage level, but I
opted for this option as it's safer.