Fix server-side errors when creating a user account using an existing handle and when authenticating using flattened params (#1161)
Fixes:
https://appsignal.com/tuist-cloud/sites/6607b64d83eb6789e61c8ba7/exceptions/incidents/445

We had the logic for this in place, but we only checked if an
organization with the same handle existed, not a user. This PR extends
the check to be across users and organizations.

Fixes:
https://appsignal.com/tuist-cloud/sites/6607b64d83eb6789e61c8ba7/exceptions/incidents/378

For some reason, in some scenarios (my guess is that is due to the
browser), the form params are flattened when included in the request.
Because we don't expect that, the authentication fails.