Not delete marketing content and sources at compile-time (#1163)
I added some logic to delete marketing content and sources at
compile-time to make on-premise images more lightweight since on-premise
customers don't need the marketing routes. However, one of the customers
ran into runtime issues because some Elixir modules could not be found.

I'm reverting that logic because it leaves the compiled graph of modules
in a weird state. It also gets tricky as the inter-dependencies between
marketing and non-marketing modules increase, which could lead to CI
passing, but things failing later on on runtime when on-premise
customers try to run the images.