Improve the generation of OG images (#1178)
I noticed the OG image for the new page was not generated. It turned out
that the task had a hardcoded list of pages that was missing this new
page. I adjusted that logic to automatically do so for the dynamic pages
like `openness` and I took the opportunity to align the name of the task
with the convention used by Phoenix mix tasks: `mix
marketing.gen.og_images`