Add support for using virtual hosts when interacting with S3 (#1185)
Documentation PR: https://github.com/tuist/tuist/pull/7124

Some on-premise customers need the bucket to be used as sub-domain when
constructing the URL (as opposed to path prefix). This is how it worked
when we used the Rust implementation of the S3 client, and that we broke
with the removal of that code in favor of an Elixir implementation. This
PR makes that configurable via the: `TUIST_S3_VIRTUAL_HOST=1` variable.

---------

Co-authored-by: Marek Fořt <marekfort@me.com>